backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.8.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
confluent-kafka[protobuf]>=2.3.0
protobuf>=6.31.1
requests>=2.31.0
supabase>=2.8.0
python-jose[cryptography]>=3.3.0
transformers>=4.41.0
torch>=2.2.0
redis>=5.0.0
python-dotenv>=1.0.0
joblib>=1.3.0
google-generativeai>=0.8.4